feat: initial commit
First commit for mill-ci-release. This is a plugin modeled after
https://github.com/sbt/sbt-ci-release, which makes publishing a breeze.
The current reasoning for this is that there seems to not be an easy way
to publish Snapshots from Mill without a lot of manual work and even the
current setup is a bit annoying to have to do for every repo. This is an
attempt to make this a bit easier for people new to Mill that just want
a quick way to publish. One liners for the win.